[Android] Centralize initialization of Profile at the Activity level
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/778674388846f118ffb841cca062989a369e40bc

Centralize initialization of Profile at the Activity level.

This introduces ProfileProvider, which hides the concept of
OTRProfileID from clients and enables swapping out the original
profile if every applicable.

TabModelSelectorProfileSupplier already exists, but that is
focused on UI state (e.g. looking at an incognito tab vs a
normal tab) and can change at runtime. ProfileProvider instead
aims to to be consistent for the lifetime of an Activity.

Low-Coverage-Reason: New classes are unit tested, activity glue can't be meaningfully covered.
Bug: 1410601,1464647