[Android] Minor color adjustments
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/ecebb05e61843b6b2b0f8dd1b66158c2a1c98583

SwitchMaterial to MaterialSwitch

Migrate the SwitchMaterial (GM2) into MaterialSwitch (GM3).

To mitigate the size increase for the MG3 switch, we are applying a 0.8 scale for the switch, while setting a bigger minHeight (60dp) in order to maintain the min touch garget size. This however brings the side effect of size could have a bigger height.

Another side effect of scaling is the text come with switch widget will be scaled too. This CL mitigate that by wrapping the switches in a linear layout and intercept the text being set in the resources.

With this migration, there are few gotchas this CL also addresses:
 - Migrate the UI using SwitchCompat into MaterialSwitch to keep the consistent look
 - Add the materialSwitchStyle to Base.Theme.BrowserUI to cooperate UIs that uses SwitchPreferenceCompat out side of SettingsActivity
 - Add themes to render tests that does not apply the theme themselves, so the switches on the goldens has the correct scale applied.

Bug: 1248590